# docs/api-documentation.yaml
openapi: 3.0.3
info:
  title: WonderWay Backend API
  description: |
    Enterprise-grade social media platform API similar to Twitter with advanced features.
    
    ## Features
    - User authentication and authorization
    - Posts, comments, likes, reposts
    - Real-time messaging and notifications
    - Live streaming capabilities
    - Monetization platform
    - Multi-language support
    
  version: 3.0.0
  contact:
    name: WonderWay Development Team
    email: dev@wonderway.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.wonderway.com/api
    description: Production server
  - url: https://staging-api.wonderway.com/api
    description: Staging server
  - url: http://localhost:8000/api
    description: Development server
security:
  - bearerAuth: []
paths:
...
  /posts:
    get:
      tags: [Posts]
      summary: Get timeline posts
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
            maximum: 50
      responses:
        '200':
          description: Timeline posts retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Post'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'

    post:
      tags: [Posts]
      summary: Create new post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
                  maxLength: 280
                  example: "Hello World! This is my first post."
                media:
                  type: array
                  items:
                    type: string
                    format: binary
                  maxItems: 4
                is_draft:
                  type: boolean
                  default: false
      responses:
        '201':
          description: Post created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        username:
          type: string
        email:
          type: string
          format: email
        avatar:
          type: string
          format: uri
        followers_count:
          type: integer
        following_count:
          type: integer
        posts_count:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Post:
      type: object
      properties:
        id:
          type: integer
        content:
          type: string
        user:
          $ref: '#/components/schemas/User'
        media:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              url:
                type: string
                format: uri
              type:
                type: string
                enum: [image, video]
        likes_count:
          type: integer
        comments_count:
          type: integer
        reposts_count:
          type: integer
        is_liked:
          type: boolean
        is_reposted:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    PostResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Post'
        message:
          type: string
    PaginationMeta:
      type: object
      properties:
        current_page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
        last_page:
          type: integer
...